Move dispatch methods directly onto app class.
The base class workaround was an artifact of my older dispatchmethods
branch, but the circular import situation that triggered it is gone now.
