+++ morepath/app.py
@@ -16,8 +16,9 @@ Entirely documented in :class:`morepath.App` in the public API.
 """
 
 import dectate
+import reg
+from webob.exc import HTTPNotFound
 
-from .generic import GenericApp
 from .request import Request
 from . import compat
 from .reify import reify
@@ -25,7 +26,11 @@ from .path import PathInfo
 from .error import LinkError
 
 
-class App(GenericApp):
+def cached_key_lookup(key_lookup):
+    return reg.CachingKeyLookup(key_lookup, 1000, 1000, 1000)
+
+
+class App(dectate.App):
     """A Morepath-based application object.
 
     You subclass App to create a morepath application class. You can
@@ -241,6 +246,151 @@ class App(GenericApp):
 
         for section, section_settings in settings.items():
             set_setting_section(section, section_settings)
+
+    @reg.dispatch_method(get_key_lookup=cached_key_lookup)
+    def _view(self, obj, request):
+        """Get the view that represents the obj in the context of a request.
+
+        This view is a representation of the obj that can be rendered to a
+        response. It may also return a :class:`morepath.Response`
+        directly.
+
+        Predicates are installed in :mod:`morepath.core` that inspect both
+        ``obj`` and ``request`` to see whether a matching view can be found.
+
+        :param obj: model object to represent with view.
+        :param request: :class:`morepath.Request` instance.
+        :return: :class:`morepath.Response` object, or
+        :class:`webob.exc.HTTPNotFound` if view cannot be found.
+        """
+        return HTTPNotFound()
+
+    @reg.dispatch_method('identity', get_key_lookup=cached_key_lookup)
+    def _verify_identity(self, identity):
+        """Returns True if the claimed identity can be verified.
+
+        Look in the database to verify the identity, or in case of auth
+        tokens, always consider known identities to be correct.
+
+        :param: :class:`morepath.Identity` instance.
+        :return: ``True`` if identity can be verified. By default no identity
+        can be verified so this returns ``False``.
+        """
+        return False
+
+    @reg.dispatch_method('identity', 'obj',
+                         reg.match_class('permission',
+                                         lambda permission: permission),
+                         get_key_lookup=cached_key_lookup)
+    def _permits(self, identity, obj, permission):
+        """Returns ``True`` if identity has permission for model object.
+
+        identity can be the special :data:`morepath.NO_IDENTITY`
+        singleton; register for :class:`morepath.NoIdentity` to handle
+        this case separately.
+
+        :param identity: :class:`morepath.Identity`
+        :param obj: model object
+        :param permission: permission class.
+        :return: ``True`` if identity has permission for obj.
+        """
+        return False
+
+    def _load_json(self, json, request):
+        """Load JSON as some object.
+
+        By default JSON is loaded as itself.
+
+        :param json: JSON (in Python form) to convert into object.
+        :param request: :class:`morepath.Request`
+        :return: Any Python object, including JSON.
+        """
+        return json
+
+    @reg.dispatch_method('obj', get_key_lookup=cached_key_lookup)
+    def _dump_json(self, obj, request):
+        """Dump an object as JSON.
+
+        ``obj`` is any Python object, try to interpret it as JSON.
+
+        :param obj: any Python object to convert to JSON.
+        :param request: :class:`morepath.Request`
+        :return: JSON representation (in Python form).
+        """
+        return obj
+
+    def _link_prefix(self, request):
+        """Returns a prefix that's added to every link generated by request.
+
+        By default :attr:`webob.request.BaseRequest.application_url` is used.
+
+        :param request: :class:`morepath.Request`
+        :return: prefix string to add before links.
+        """
+        return request.application_url
+
+    @reg.dispatch_method(reg.match_class('model', lambda model: model),
+                         get_key_lookup=cached_key_lookup)
+    def _class_path(self, model, variables):
+        """Get the path for a model class.
+
+        :param model: model class or :class:`morepath.App` subclass.
+        :param variables: dictionary with variables to reconstruct
+        the path and URL paramaters from path pattern.
+        :return: a :class:`morepath.path.PathInfo` with path within this app,
+          or ``None`` if the path couldn't be determined.
+        """
+        return None
+
+    @reg.dispatch_method('obj', get_key_lookup=cached_key_lookup)
+    def _path_variables(self, obj):
+        """Get variables to use in path generation.
+
+        :param obj: model object or :class:`morepath.App` instance.
+        :return: a dict with the variables to use for constructing the path,
+        or ``None`` if no such dict can be found.
+        """
+        return self._default_path_variables(obj)
+
+    @reg.dispatch_method('obj', get_key_lookup=cached_key_lookup)
+    def _default_path_variables(self, obj):
+        """Get default variables to use in path generation.
+
+        Invoked if no specific ``path_variables`` is registered.
+
+        :param obj: model object for ::class:`morepath.App` instance.
+        :return: a dict with the variables to use for constructing the
+        path, or ``None`` if no such dict can be found.
+        """
+        return None
+
+    @reg.dispatch_method('obj', get_key_lookup=cached_key_lookup)
+    def _deferred_link_app(self, obj):
+        """Get application used for link generation.
+
+        :param obj: model object to link to.
+        :return: instance of :class:`morepath.App` subclass that handles
+        link generation for this model, or ``None`` if no app exists
+        that can construct link.
+        """
+        return None
+
+    @reg.dispatch_method(reg.match_class('model', lambda model: model),
+                         get_key_lookup=cached_key_lookup)
+    def _deferred_class_link_app(self, model, variables):
+        """Get application used for link generation for a model class.
+
+        :param model: model class
+        :param variables: dict of variables used to construct class link
+        :return: instance of :class:`morepath.App` subclass that handles
+        link generation for this model class, or ``None`` if no app exists
+        that can construct link.
+        """
+        return None
+
+    @classmethod
+    def clean(cls):
+        reg.clean_dispatch_methods(cls)
 
     def _identify(self, request):
         """Determine identity for request.
